requirements.txt
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Deprecated==1.2.14
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.2
toml==0.10.2
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
